fix sorttable import and navbar back button styles
